+++ components/Analysis/AnalysisSelectors/AnalysisDecimationSelector.tsx
