+++ components/Analysis/AnalysisSelectors/AnalysisViewSelector.tsx
